tests/wpt/tests/html/browsers/history/the-location-interface/scripted_form_submit_assign_during_load.html

<!doctype html>
<title>Assignment to location with form submit during load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p>The popup blocker must be disabled for this test</p>
<div id="log"></div>
<script>
setup({timeout:3600000});
var t = async_test();
var win = window.open("scripted_form_submit_assign_during_load-1.html");

var history_length;
do_test = t.step_func(function(new_length) {
  assert_equals(new_length, history_length);
  t.done();
});
</script>
